Made DB build-up tear-down inot a mixin
