+++ .idea/libraries/SBT__org_webjars_jquery_1_11_1_jar.xml
@@ -1,0 +1,11 @@
+<component name="libraryTable">
+  <library name="SBT: org.webjars:jquery:1.11.1:jar">
+    <CLASSES>
+      <root url="jar://$USER_HOME$/.ivy2/cache/org.webjars/jquery/jars/jquery-1.11.1.jar!/" />
+    </CLASSES>
+    <JAVADOC />
+    <SOURCES>
+      <root url="jar://$USER_HOME$/.ivy2/cache/org.webjars/jquery/srcs/jquery-1.11.1-sources.jar!/" />
+    </SOURCES>
+  </library>
+</component>
