+++ fijbook-alpha/test/com/fijimf/deepfij/models/RepoGamePredictionSpec.scala
@@ -15,20 +15,12 @@ import scala.concurrent.duration.{Duration, _}
 import scala.util.{Failure, Success}
 
 
-class RepoGamePredictionSpec extends PlaySpec with OneAppPerTest with BeforeAndAfterEach with ScalaFutures {
+class RepoGamePredictionSpec extends PlaySpec with OneAppPerTest with BeforeAndAfterEach  with RebuildDatabaseMixin with ScalaFutures {
   implicit override val patienceConfig = PatienceConfig(timeout = Span(3, Seconds), interval = Span(250, Millis))
-  val repo = Injector.inject[ScheduleRepository]
   val dao = Injector.inject[ScheduleDAO]
 
   import scala.concurrent.ExecutionContext.Implicits.global
 
-  override def beforeEach() = {
-    Await.result(repo.createSchema(), Duration.Inf)
-  }
-
-  override def afterEach() = {
-    Await.result(repo.dropSchema(), Duration.Inf)
-  }
 
   /*
