+++ fijbook-alpha/test/com/fijimf/deepfij/models/RepoLogisticModelSpec.scala
@@ -15,18 +15,11 @@ import scala.concurrent.{Await, Future}
 import scala.util.{Failure, Success}
 
 
-class RepoLogisticModelSpec extends PlaySpec with OneAppPerTest with BeforeAndAfterEach with ScalaFutures {
+class RepoLogisticModelSpec extends PlaySpec with OneAppPerTest with BeforeAndAfterEach with RebuildDatabaseMixin  with ScalaFutures {
   implicit override val patienceConfig = PatienceConfig(timeout = Span(3, Seconds), interval = Span(250, Millis))
-  val repo = Injector.inject[ScheduleRepository]
   val dao = Injector.inject[ScheduleDAO]
 
-  override def beforeEach() = {
-    Await.result(repo.createSchema(), Duration.Inf)
-  }
 
-  override def afterEach() = {
-    Await.result(repo.dropSchema(), Duration.Inf)
-  }
 /*
 
   def listLogisticModelParameters: Future[List[LogisticModelParameter]]
