+++ fijbook-alpha/test/com/fijimf/deepfij/models/RepoTeamSpec.scala
@@ -14,19 +14,10 @@ import scala.concurrent.duration.Duration
 import scala.concurrent.{Await, Future}
 
 
-class RepoTeamSpec extends PlaySpec with OneAppPerTest with BeforeAndAfterEach with ScalaFutures {
+class RepoTeamSpec extends PlaySpec with OneAppPerTest with BeforeAndAfterEach with RebuildDatabaseMixin with ScalaFutures {
   implicit override val patienceConfig = PatienceConfig(timeout = Span(3, Seconds), interval = Span(250, Millis))
-  val repo = Injector.inject[ScheduleRepository]
+
   val dao = Injector.inject[ScheduleDAO]
-
-  override def beforeEach() = {
-    Await.result(repo.createSchema(), Duration.Inf)
-  }
-
-  override def afterEach() = {
-    Await.result(repo.dropSchema(), Duration.Inf)
-  }
-
 
   "Teams " should {
     "be empty initially" in new WithApplication(FakeApplication()) {
